Move some vtop steps to code freeze
