+++ go/releaser/code_freeze/vtop_update_compatibility_table.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package pre_release
+package code_freeze
 
 import (
 	"fmt"
